# openapi/users-swagger2.yaml
swagger: "2.0"
info:
  title: LionSwap Identity & Accounts API
  version: "0.1.0"
  description: Columbia LionMail-authenticated user identities

# --- IMPORTANT FOR CLOUD RUN ---
host: ms1-identity-157498364441.us-east1.run.app
basePath: /
schemes:
  - https
  - http


consumes:
  - application/json
produces:
  - application/json

paths:
  /users:
    get:
      summary: List users
      tags: [users]
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/User"

    post:
      summary: Create user
      tags: [users]
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UserCreate"
      responses:
        201:
          description: Created
          headers:
            Location:
              type: string
              description: URI of the created resource
          schema:
            $ref: "#/definitions/User"

  /users/{uni}:
    parameters:
      - name: uni
        in: path
        required: true
        type: string

    get:
      summary: Get user by uni
      tags: [users]
      parameters:
        - in: header
          name: If-None-Match
          type: string
          required: false
      responses:
        200:
          description: OK
          headers:
            ETag:
              type: string
          schema:
            $ref: "#/definitions/User"
        304:
          description: Not Modified
        404:
          description: Not Found

    put:
      summary: Replace user info with uni
      tags: [users]
      parameters:
        - in: header
          name: If-Match
          type: string
          required: true
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UserUpdate"
      responses:
        200:
          description: Updated
          headers:
            ETag:
              type: string
          schema:
            $ref: "#/definitions/User"
        412:
          description: ETag mismatch
        428:
          description: If-Match required
        404:
          description: Not Found

    delete:
      summary: Delete user with uni
      tags: [users]
      responses:
        200:
          description: Deleted
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Successfully deleted "ab1234" user'
        404:
          description: Not Found

  /users/{uni}/profile:
    parameters:
      - name: uni
        in: path
        required: true
        type: string
    get:
      summary: Get user profile (relative path)
      description: |
        Relative path endpoint demonstrating HATEOAS principles.
        Returns a subset of user data focused on profile information.
        This endpoint follows the pattern: /users/{uni}/profile
      tags: [users]
      parameters:
        - in: header
          name: If-None-Match
          type: string
          required: false
      responses:
        200:
          description: OK
          headers:
            ETag:
              type: string
          schema:
            $ref: "#/definitions/UserProfile"
        304:
          description: Not Modified
        404:
          description: Not Found

definitions:
  User:
    type: object
    required: [uni, student_name, email]
    properties:
      user_id: { type: integer, format: int64, example: 1 }
      uni: { type: string, example: "ab1234" }
      student_name: { type: string, example: "Sally Go" }
      dept_name: { type: string, example: "CS" }
      email: { type: string, format: email, example: "sg####@columbia.edu" }
      phone: { type: string, nullable: true }
      avatar_url: { type: string, nullable: true }
      credibility_score: { type: number, format: float, example: 0.0 }
      last_seen_at: { type: string, format: date-time, nullable: true }
      _links:
        type: object
        description: HATEOAS links for navigation
        properties:
          self:
            $ref: "#/definitions/Link"
          profile:
            $ref: "#/definitions/Link"
          update:
            allOf:
              - $ref: "#/definitions/Link"
              - type: object
                properties:
                  method:
                    type: string
                    example: "PUT"
          delete:
            allOf:
              - $ref: "#/definitions/Link"
              - type: object
                properties:
                  method:
                    type: string
                    example: "DELETE"

  UserCreate:
    type: object
    required: [uni, student_name, email]
    properties:
      uni: { type: string }
      student_name: { type: string }
      dept_name: { type: string }
      email: { type: string, format: email }
      phone: { type: string }
      avatar_url: { type: string }

  UserUpdate:
    type: object
    properties:
      student_name: { type: string }
      dept_name: { type: string }
      phone: { type: string }

  UserProfile:
    type: object
    description: User profile information (relative path response)
    properties:
      uni: { type: string, example: "ab1234" }
      student_name: { type: string, example: "Sally Go" }
      dept_name: { type: string, example: "CS" }
      avatar_url: { type: string, nullable: true }
      credibility_score: { type: number, format: float, example: 0.0 }
      last_seen_at: { type: string, format: date-time, nullable: true }
      _links:
        type: object
        description: HATEOAS links for navigation
        properties:
          self:
            $ref: "#/definitions/Link"
          user:
            $ref: "#/definitions/Link"
          update:
            allOf:
              - $ref: "#/definitions/Link"
              - type: object
                properties:
                  method:
                    type: string
                    example: "PUT"

  Link:
    type: object
    description: HATEOAS link object
    properties:
      href: { type: string, description: "Relative path to the resource" }
      method: { type: string, description: "HTTP method (for action links)", required: false }
    required: [href]

  Links:
    type: object
    description: Collection of HATEOAS links
    additionalProperties:
      $ref: "#/definitions/Link"